chore: Remove default config of database
If someone clone or fork this project they need create a database
by yourself on Firebase because I am removing this database.
But the online app is continue working with another config.

This config information was there because the reviewer of the project
had needed to run without create a new database.